fix: circular dependencies in WebWorker
