+++ src/WebWorker.ts
@@ -5,11 +5,3 @@ export const createUploadWebWorker = (): Worker => {
 export const createDownloadWebWorker = (): Worker => {
   return new Worker(new URL('./download.worker', import.meta.url), { type: 'module' });
 };
-
-export const WORKER_MESSAGE_STATES = {
-  SUCCESS: 'success',
-  ERROR: 'error',
-  ABORT: 'abort',
-  CHECK_UPLOAD_STATUS: 'checkUploadStatus',
-  UPLOAD_STATUS: 'uploadStatus',
-};
