+++ src/app/network/NetworkFacade.ts
@@ -8,9 +8,8 @@ import { EncryptFileFunction, UploadFileMultipartFunction } from '@internxt/sdk/
 import envService from 'app/core/services/env.service';
 import { buildProgressStream } from 'app/core/services/stream.service';
 import { queue, QueueObject } from 'async';
-import { WORKER_MESSAGE_STATES } from '../../WebWorker';
 
-import { waitForContinueUploadSignal } from '../drive/services/worker.service/uploadWorkerUtils';
+import { waitForContinueUploadSignal, WORKER_MESSAGE_STATES } from '../drive/services/worker.service/uploadWorkerUtils';
 import { TaskStatus } from '../tasks/types';
 import { encryptStreamInParts, generateFileKey, getEncryptedFile, processEveryFileBlobReturnHash } from './crypto';
 import { DownloadProgressCallback, getDecryptedStream } from './download';
